authoz

+++ Controllers/CarController.cs
@@ -9,9 +9,10 @@ using TemplateWebApiPhucThinh.Data.Model;
 
 using TemplateWebApiPhucThinh.ModelValidation;
 using TemplateWebApiPhucThinh.Repository.IRepository;
-
+using Microsoft.AspNetCore.Authorization;
 namespace TemplateWebApiPhucThinh.Controllers
 {
+    [Authorize()]
     [Route("[controller]")]
     [ApiController]
     public class CarController : ControllerBase
@@ -30,6 +31,13 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("Create")]
         public IActionResult Create([FromBody] Car Car)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
+      
             Car.Id = Guid.NewGuid() + "";
             Car.IsDelete=false;
             return Ok(_repository.Create(Car));
@@ -40,6 +48,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("GetById/{id}")]
         public IActionResult GetById(string id)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
             return Ok(_repository.GetById(id));
         }
 
@@ -47,19 +61,36 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("Delete/{id}")]
         public IActionResult Delete(string id)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
             return Ok(_repository.Delete(id));
         }
         [HttpPut]
         [Route("Update/{id}")]
         public IActionResult Update(string id, [FromBody] Car Car)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
             return Ok(_repository.Update(id, Car));
         }
         [HttpGet]
         [Route("Paging/pagesize/pageNow")]
         public IActionResult Paging(int pagesize, int pageNow)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+                return Forbid();
+            }
             
             return Ok(_repository.Paging(pagesize, pageNow, "color"));
 
@@ -69,7 +100,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("CountOfPaging/pagesize/pageNow")]
         public IActionResult CountOfPaging(int pagesize, int pageNow)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+                return Forbid();
+            }
             
             return Ok(_repository.CountOfPaging(pagesize, pageNow));
 
@@ -78,7 +114,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("CountAll/pagesize/pageNow")]
         public IActionResult CountAll(int pagesize, int pageNow)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+                return Forbid();
+            }
             
             return Ok(_repository.CountAll());
 
@@ -87,6 +128,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("DeleteEnable/{id}")]
         public IActionResult DeleteEnable(string id)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
             if (string.IsNullOrWhiteSpace(id))
             {
                 return BadRequest();
@@ -98,7 +145,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("PagingCondition/pagesize/pageNow/condition")]
         public IActionResult PagingCondition(int pagesize, int pageNow,string condition)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+                return Forbid();
+            }
             
             return Ok(_repository.PagingCondition(pagesize, pageNow, condition));
 
@@ -108,6 +160,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("CountCondition/condition")]
         public IActionResult CountCondition(string condition)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Forbid();
+            }
             return Ok(_repository.CountCondition(condition));
         }
 
@@ -116,7 +174,12 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("PagingConditionPrice/condition/pageIndex/pageSize/sortOrder/priceStart/priceEnd")]
         public IActionResult PagingConditionPrice(string condition, int pageIndex, int pageSize, string sortOrder, int priceStart, int priceEnd)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("name")){
+            if( claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+                return Forbid();
+            }
             
             return Ok(_repository.PagingConditionPrice( condition,  pageIndex,  pageSize,  sortOrder,  priceStart,  priceEnd));
