+++ Controllers/PartnerController.cs
@@ -9,9 +9,10 @@ using TemplateWebApiPhucThinh.Data.Model;
 
 using TemplateWebApiPhucThinh.ModelValidation;
 using TemplateWebApiPhucThinh.Repository.IRepository;
-
+using Microsoft.AspNetCore.Authorization;
 namespace TemplateWebApiPhucThinh.Controllers
 {
+    [Authorize()]
     [Route("[controller]")]
     [ApiController]
     public class PartnerController : ControllerBase
@@ -30,9 +31,20 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("Create")]
         public IActionResult Create([FromBody] Partner Partner)
         {
-            Partner.Id = Guid.NewGuid() + "";
-            Partner.IsDelete=false;
-            return Ok(_repository.Create(Partner));
+        var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com")|| claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                Partner.Id = Guid.NewGuid() + "";
+                Partner.IsDelete=false;
+                return Ok(_repository.Create(Partner));
+            }
+            else
+            {
+                return Forbid();
+            }
+      
+           
         }
      
 
@@ -40,28 +52,71 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("GetById/{id}")]
         public IActionResult GetById(string id)
         {
-            return Ok(_repository.GetById(id));
+            var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com")|| claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+               return Ok(_repository.GetById(id));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
+          
         }
 
         [HttpDelete]
         [Route("Delete/{id}")]
         public IActionResult Delete(string id)
         {
-            return Ok(_repository.Delete(id));
+              var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com")|| claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+              return Ok(_repository.Delete(id));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
+           
         }
         [HttpPut]
         [Route("Update/{id}")]
         public IActionResult Update(string id, [FromBody] Partner Partner)
         {
-            return Ok(_repository.Update(id, Partner));
+              var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+             return Ok(_repository.Update(id, Partner));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
+           
         }
         [HttpGet]
         [Route("Paging/pagesize/pageNow")]
         public IActionResult Paging(int pagesize, int pageNow)
         {
             
+            var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
-            return Ok(_repository.Paging(pagesize, pageNow, "color"));
+                return Ok(_repository.Paging(pagesize, pageNow, "color"));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
+            
 
         }
 
@@ -69,38 +124,80 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("CountOfPaging/pagesize/pageNow")]
         public IActionResult CountOfPaging(int pagesize, int pageNow)
         {
+             var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+                return Ok(_repository.CountOfPaging(pagesize, pageNow));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
             
             
-            return Ok(_repository.CountOfPaging(pagesize, pageNow));
+           
 
         }
          [HttpGet]
         [Route("CountAll/pagesize/pageNow")]
         public IActionResult CountAll(int pagesize, int pageNow)
         {
+              var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+              return Ok(_repository.CountAll());
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
             
-            return Ok(_repository.CountAll());
+           
 
         }
         [HttpDelete]
         [Route("DeleteEnable/{id}")]
         public IActionResult DeleteEnable(string id)
         {
-            if (string.IsNullOrWhiteSpace(id))
+               var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+               if (string.IsNullOrWhiteSpace(id))
             {
                 return BadRequest();
             }
             return Ok(_repository.DeleteEnable(id));
+            }
+            else
+            {
+                return Forbid();
+            }
+           
+          
         }
 
         [HttpGet]
         [Route("PagingCondition/pagesize/pageNow/condition")]
         public IActionResult PagingCondition(int pagesize, int pageNow,string condition)
         {
+              var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+              return Ok(_repository.PagingCondition(pagesize, pageNow, condition));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
             
-            return Ok(_repository.PagingCondition(pagesize, pageNow, condition));
+          
 
         }
 
@@ -108,7 +205,18 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("CountCondition/condition")]
         public IActionResult CountCondition(string condition)
         {
-            return Ok(_repository.CountCondition(condition));
+            var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
+            
+               return Ok(_repository.CountCondition(condition));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
+           
         }
 
 
@@ -116,9 +224,19 @@ namespace TemplateWebApiPhucThinh.Controllers
         [Route("PagingConditionPrice/condition/pageIndex/pageSize/sortOrder/priceStart/priceEnd")]
         public IActionResult PagingConditionPrice(string condition, int pageIndex, int pageSize, string sortOrder, int priceStart, int priceEnd)
         {
+                   var claims = User.Claims.Select(claim => new { claim.Type, claim.Value }).ToDictionary( t => t.Type, t => t.Value);
+            if(claims.ContainsKey("email")){
+            if(claims["email"].Equals("tmdt2019@gmail.com") || claims["name"].Equals("NhanVienTMDT NhanVienTMDT"))
             
+               return Ok(_repository.PagingConditionPrice( condition,  pageIndex,  pageSize,  sortOrder,  priceStart,  priceEnd));
+            }
+            else
+            {
+                return Forbid();
+            }
+            return Forbid();
             
-            return Ok(_repository.PagingConditionPrice( condition,  pageIndex,  pageSize,  sortOrder,  priceStart,  priceEnd));
+           
 
         }
     }
